+++ sdk/storage/Azure.Storage.Files.DataLake/api/Azure.Storage.Files.DataLake.netstandard2.0.cs
@@ -17,6 +17,8 @@ namespace Azure.Storage.Files.DataLake
     public partial class DataLakeDirectoryClient : Azure.Storage.Files.DataLake.DataLakePathClient
     {
         protected DataLakeDirectoryClient() { }
+        public DataLakeDirectoryClient(string connectionString, string fileSystemName, string directoryPath) { }
+        public DataLakeDirectoryClient(string connectionString, string fileSystemName, string directoryPath, Azure.Storage.Files.DataLake.DataLakeClientOptions options) { }
         public DataLakeDirectoryClient(System.Uri directoryUri) { }
         public DataLakeDirectoryClient(System.Uri directoryUri, Azure.Core.TokenCredential credential) { }
         public DataLakeDirectoryClient(System.Uri directoryUri, Azure.Core.TokenCredential credential, Azure.Storage.Files.DataLake.DataLakeClientOptions options) { }
@@ -63,6 +65,8 @@ namespace Azure.Storage.Files.DataLake
     public partial class DataLakeFileClient : Azure.Storage.Files.DataLake.DataLakePathClient
     {
         protected DataLakeFileClient() { }
+        public DataLakeFileClient(string connectionString, string fileSystemName, string filePath) { }
+        public DataLakeFileClient(string connectionString, string fileSystemName, string filePath, Azure.Storage.Files.DataLake.DataLakeClientOptions options) { }
         public DataLakeFileClient(System.Uri fileUri) { }
         public DataLakeFileClient(System.Uri fileUri, Azure.Core.TokenCredential credential) { }
         public DataLakeFileClient(System.Uri fileUri, Azure.Core.TokenCredential credential, Azure.Storage.Files.DataLake.DataLakeClientOptions options) { }
@@ -148,6 +152,8 @@ namespace Azure.Storage.Files.DataLake
     public partial class DataLakeFileSystemClient
     {
         protected DataLakeFileSystemClient() { }
+        public DataLakeFileSystemClient(string connectionString, string fileSystemName) { }
+        public DataLakeFileSystemClient(string connectionString, string fileSystemName, Azure.Storage.Files.DataLake.DataLakeClientOptions options) { }
         public DataLakeFileSystemClient(System.Uri fileSystemUri) { }
         public DataLakeFileSystemClient(System.Uri fileSystemUri, Azure.Core.TokenCredential credential) { }
         public DataLakeFileSystemClient(System.Uri fileSystemUri, Azure.Core.TokenCredential credential, Azure.Storage.Files.DataLake.DataLakeClientOptions options) { }
@@ -219,6 +225,8 @@ namespace Azure.Storage.Files.DataLake
     {
         protected DataLakePathClient() { }
         public DataLakePathClient(Azure.Storage.Files.DataLake.DataLakeFileSystemClient fileSystemClient, string path) { }
+        public DataLakePathClient(string connectionString, string fileSystemName, string path) { }
+        public DataLakePathClient(string connectionString, string fileSystemName, string path, Azure.Storage.Files.DataLake.DataLakeClientOptions options) { }
         public DataLakePathClient(System.Uri pathUri) { }
         public DataLakePathClient(System.Uri pathUri, Azure.Core.TokenCredential credential) { }
         public DataLakePathClient(System.Uri pathUri, Azure.Core.TokenCredential credential, Azure.Storage.Files.DataLake.DataLakeClientOptions options) { }
@@ -267,6 +275,8 @@ namespace Azure.Storage.Files.DataLake
     public partial class DataLakeServiceClient
     {
         protected DataLakeServiceClient() { }
+        public DataLakeServiceClient(string connectionString) { }
+        public DataLakeServiceClient(string connectionString, Azure.Storage.Files.DataLake.DataLakeClientOptions options) { }
         public DataLakeServiceClient(System.Uri serviceUri) { }
         public DataLakeServiceClient(System.Uri serviceUri, Azure.Core.TokenCredential credential) { }
         public DataLakeServiceClient(System.Uri serviceUri, Azure.Core.TokenCredential credential, Azure.Storage.Files.DataLake.DataLakeClientOptions options) { }